Sushil: Penalized LS working. Checked against Boyd's L1 toolbox.
git-svn-id: http://bsccs.googlecode.com/svn/branches/least_squares@88 f7c91f03-731f-d806-b965-e12744d8cf89

+++ C/codebase/CCD/CCTestInputReader.cpp
@@ -92,7 +92,7 @@ void CCTestInputReader::readFile(const char* fileName) {
 			pid.push_back(numCases - 1);
 
 			// Parse outcome entry
-			int thisEta;
+			double thisEta;
 			istringstream(strVector[colOutcome]) >> thisEta;
  			numEvents += thisEta;
 			eta.push_back(thisEta);
@@ -107,7 +107,6 @@ void CCTestInputReader::readFile(const char* fileName) {
 				if (value != 0) {
 					value = 1;
 				}
-
 			}
 			add_data(0, currentRow, value);
